requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
attrs==17.4.0
autopep8==1.3.5
Babel==2.5.3
billiard==3.5.0.3
bleach==2.1.3
boto3==1.5.31
botocore==1.8.45
braintree==3.39.1
cairocffi==0.8.0
CairoSVG==2.1.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.1.4
cssselect2==0.2.1
custom-sastrawi==1.0.5
defusedxml==0.5.0
dj-database-url==0.4.2
dj-email-url==0.0.10
Django==2.0.3
django-babel==0.6.2
django-bootstrap4==0.0.5
django-cache-url==2.0.0
django-celery-results==1.0.1
django-countries==5.2
django-debug-toolbar==1.9.1
django-elasticsearch-dsl==0.4.4
django-environ==0.4.4
django-filter==1.1.0
django-fsm==2.6.0
django-graphql-jwt==0.1.9
django-impersonate==1.3
django-js-asset==1.0.0
django-mptt==0.9.0
django-payments==0.12.1
django-phonenumber-field==2.0.0
django-prices==1.0.0
django-prices-openexchangerates==1.0.0
django-redis==4.8.0
django-render-block==0.5
django-storages==1.6.5
django-templated-email==2.2.0
django-versatileimagefield==1.9
django-webpack-loader==0.5.0
djangorestframework==3.8.2
docutils==0.14
elasticsearch==5.5.2
elasticsearch-dsl==5.4.0
Faker==0.8.11
first==2.0.1
freezegun==0.3.9
google-i18n-address==2.3.0
google-measurement-protocol==1.0.0
gprof2dot==2016.10.13
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.6
iso8601==0.1.12
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.0
jsonfield==2.0.2
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
maxminddb==1.3.0
maxminddb-geolite2==2017.803
multidict==3.3.2
oauthlib==2.0.6
pdfrw==0.4
phonenumberslite==8.9.1
Pillow==5.0.0
pip-tools==2.0.2
pluggy==0.6.0
prices==1.0.0
promise==2.1
psycopg2-binary==2.7.4
purl==1.3.1
py==1.5.2
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
Pyphen==0.9.4
pytest==3.4.0
pytest-django==3.1.2
pytest-mock==1.7.0
pytest-vcr==0.3.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
Rx==1.6.0
s3transfer==0.1.12
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
sqlparse==0.2.4
stripe==1.77.1
suds-jurko==0.6
text-unidecode==1.1
tinycss2==0.6.1
typing==3.6.2
urllib3==1.22
vcrpy==1.11.1
vine==1.1.4
WeasyPrint==0.42.2
webencodings==0.5.1
wrapt==1.10.11
xmltodict==0.11.0
yarl==0.17.0
